Fixes to make module work on other systems
Ignore-this: 17098647159a4c6055a03174961a30ef

darcs-hash:20110816190646-82ea9-fa459c18e49a36d10d3ab50c7dd2cd53a2950cdb.gz

+++ cmodels.c
@@ -600,7 +600,7 @@ double mutual_information(Graph_t G0, Graph_t G1) {
 						  n0, n1);
       if (n_shared == 0)
 	continue;
-      MI += (n_shared/(float)N) * log2(n_shared*N/((float)n0*n1));
+      MI += (n_shared/(float)N) * log2(n_shared*N/((double)n0*n1));
     }
   }
   return (MI);
